requirements.txt
beautifulsoup4==4.12.2
Pillow==10.0.0
pypinyin==0.49.0
Requests==2.31.0
Unidecode==1.3.6
